+++ client/src/components/Banner/LargeBanner.vue
