+++ client/src/components/Banner/MediumBanner.vue
